+++ ui/src/main/java/com/kotcrab/vis/ui/widget/MultiSplitPane.java
@@ -39,6 +39,8 @@ import com.kotcrab.vis.ui.FocusManager;
 import com.kotcrab.vis.ui.VisUI;
 import com.kotcrab.vis.ui.util.CursorManager;
 
+import java.util.Arrays;
+
 /**
  * Similar to {@link VisSplitPane} but supports multiple widgets with multiple split bars at once. Use {@link #setWidgets(Actor...)}
  * after creating to set pane widgets.
@@ -107,13 +109,25 @@ public class MultiSplitPane extends WidgetGroup {
 						currentCursor = targetCursor;
 					}
 				} else {
-					if (currentCursor != null) {
-						CursorManager.restoreDefaultCursor();
-						currentCursor = null;
-					}
+					clearCustomCursor();
 				}
 
 				return false;
+			}
+
+			@Override
+			public void exit (InputEvent event, float x, float y, int pointer, Actor toActor) {
+				super.exit(event, x, y, pointer, toActor);
+				if (pointer == -1) {
+					clearCustomCursor();
+				}
+			}
+
+			private void clearCustomCursor () {
+				if (currentCursor != null) {
+					CursorManager.restoreDefaultCursor();
+					currentCursor = null;
+				}
 			}
 		});
 
@@ -346,6 +360,11 @@ public class MultiSplitPane extends WidgetGroup {
 
 	/** Changes widgets of this split pane. You can pass any number of actors even 1 or 0. Actors can't be null. */
 	public void setWidgets (Actor... actors) {
+		setWidgets(Arrays.asList(actors));
+	}
+
+	/** Changes widgets of this split pane. You can pass any number of actors even 1 or 0. Actors can't be null. */
+	public void setWidgets (Iterable<Actor> actors) {
 		clearChildren();
 		widgetBounds.clear();
 		scissors.clear();
@@ -358,8 +377,8 @@ public class MultiSplitPane extends WidgetGroup {
 			scissors.add(new Rectangle());
 		}
 		float currentSplit = 0;
-		float splitAdvance = 1f / actors.length;
-		for (int i = 0; i < actors.length - 1; i++) {
+		float splitAdvance = 1f / getChildren().size;
+		for (int i = 0; i < getChildren().size - 1; i++) {
 			handleBounds.add(new Rectangle());
 			currentSplit += splitAdvance;
 			splits.add(currentSplit);
@@ -375,8 +394,8 @@ public class MultiSplitPane extends WidgetGroup {
 	public void setSplit (int handleBarIndex, float split) {
 		if (handleBarIndex < 0) throw new IllegalStateException("handleBarIndex can't be < 0");
 		if (handleBarIndex >= splits.size) throw new IllegalStateException("handleBarIndex can't be >= splits size");
-		float minSplit = handleOverIndex == 0 ? 0 : splits.get(handleOverIndex - 1);
-		float maxSplit = handleOverIndex == splits.size - 1 ? 1 : splits.get(handleOverIndex + 1);
+		float minSplit = handleBarIndex == 0 ? 0 : splits.get(handleBarIndex - 1);
+		float maxSplit = handleBarIndex == splits.size - 1 ? 1 : splits.get(handleBarIndex + 1);
 		split = MathUtils.clamp(split, minSplit, maxSplit);
 		splits.set(handleBarIndex, split);
 	}